Module Contract.For_RPC

Functions used exclusively for RPC calls

val get_staked_balance : context -> t -> Tez.t option Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val get_unstaked_frozen_balance : context -> t -> Tez.t option Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val get_unstaked_finalizable_balance : context -> t -> Tez.t option Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val get_full_balance : context -> t -> Tez.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val get_estimated_own_pending_slashed_amount : context -> t -> Tez.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

get_estimated_own_pending_slashed_amount ctxt contract returns the estimated own pending slashed amount of the given contract according to the currently available denunciations.