Module Rewards.For_RPC

type reward_kind =

| Baking_reward_fixed_portion

| Baking_reward_bonus_per_slot

| Attesting_reward_per_slot

| Seed_nonce_revelation_tip

| Vdf_revelation_tip

val reward_from_constants : ?coeff:Tezos_protocol_environment_019_PtParisB.Q.t -> Constants.Parametric.t -> reward_kind:reward_kind -> Tez.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

reward_from_constants ~coeff csts ~reward_kind returns the amount of rewards in Tez.t for the given reward_kind, according to the given parameters in csts. The (optional) value coeff is a multiplicative factor applied to the rewards (default = 1). It verifies reward_from_constants ~coeff csts ~reward_kind = coeff * reward_from_constants csts ~reward_kind.

val liquidity_baking_subsidy_from_constants : Constants.Parametric.t -> Tez.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

val get_reward_coeff : t -> cycle:Cycle.t -> Tezos_protocol_environment_019_PtParisB.Q.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

get_reward_coeff ctxt cycle reads the reward coeff for the given cycle from the storage. Returns Q.one if the given cycle is not between current_cycle and current_cycle + consensus_rights_delay. If adaptive issuance has not been activated, or has been activated and the given cycle is less than consensus_rights_delay after the activation cycle, then this function returns Q.one. Used only for RPCs. To get the actual rewards, use the reward functions defined above.

val get_reward_bonus : t -> cycle:Cycle.t option -> Issuance_bonus_repr.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

get_reward_bonus ctxt cycle reads the reward bonus for the given cycle from the storage. If cycle is None, returns 0.

Returns 0 if the given cycle is not between current_cycle and current_cycle + consensus_rights_delay.

If adaptive issuance has not been activated, then this function returns 0. Used only for RPCs. To get the actual rewards, use Delegate_rewards.