Module Tezos_raw_protocol_019_PtParisB.Block_header_repr

Representation of block headers.

type contents = {

payload_hash : Block_payload_hash.t;

payload_round : Round_repr.t;

seed_nonce_hash : Nonce_hash.t option;

proof_of_work_nonce : bytes;

per_block_votes : Per_block_votes_repr.per_block_votes;

}

type protocol_data = {

contents : contents;

signature : Tezos_protocol_environment_019_PtParisB.Signature.t;

}

type t = {

shell : Tezos_protocol_environment_019_PtParisB.Block_header.shell_header;

protocol_data : protocol_data;

}

type block_header = t

type raw = Tezos_protocol_environment_019_PtParisB.Block_header.t

type shell_header = Tezos_protocol_environment_019_PtParisB.Block_header.shell_header

val raw : block_header -> raw

val encoding : block_header Tezos_protocol_environment_019_PtParisB.Data_encoding.encoding

val raw_encoding : raw Tezos_protocol_environment_019_PtParisB.Data_encoding.t

val contents_encoding : contents Tezos_protocol_environment_019_PtParisB.Data_encoding.t

val unsigned_encoding : (Tezos_protocol_environment_019_PtParisB.Block_header.shell_header * contents) Tezos_protocol_environment_019_PtParisB.Data_encoding.t

val protocol_data_encoding : protocol_data Tezos_protocol_environment_019_PtParisB.Data_encoding.encoding

val shell_header_encoding : shell_header Tezos_protocol_environment_019_PtParisB.Data_encoding.encoding

type block_watermark =

| Block_header of Tezos_protocol_environment_019_PtParisB.Chain_id.t

val to_watermark : block_watermark -> Tezos_protocol_environment_019_PtParisB.Signature.watermark

val of_watermark : Tezos_protocol_environment_019_PtParisB.Signature.watermark -> block_watermark option

val max_header_length : int

The maximum size of block headers in bytes

val hash : block_header -> Tezos_protocol_environment_019_PtParisB.Block_hash.t

val hash_raw : raw -> Tezos_protocol_environment_019_PtParisB.Block_hash.t

type Tezos_protocol_environment_019_PtParisB.Error_monad.error +=

| Invalid_stamp

module Proof_of_work : sig ... end

Checks if the header that would be built from the given components is valid for the given difficulty. The signature is not passed as it is does not impact the proof-of-work stamp. The stamp is checked on the hash of a block header whose signature has been zeroed-out.

val check_timestamp : Round_repr.Durations.t -> timestamp:Tezos_protocol_environment_019_PtParisB.Time.t -> round:Round_repr.t -> predecessor_timestamp:Tezos_protocol_environment_019_PtParisB.Time.t -> predecessor_round:Round_repr.t -> unit Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

check_timestamp ctxt timestamp round predecessor_timestamp predecessor_round verifies that the block's timestamp and round are coherent with the predecessor block's timestamp and round. Fails with an error if that is not the case.

val check_signature : t -> Tezos_protocol_environment_019_PtParisB.Chain_id.t -> Tezos_protocol_environment_019_PtParisB.Signature.Public_key.t -> unit Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

val begin_validate_block_header : block_header:t -> chain_id:Tezos_protocol_environment_019_PtParisB.Chain_id.t -> predecessor_timestamp:Tezos_protocol_environment_019_PtParisB.Time.t -> predecessor_round:Round_repr.t -> fitness:Fitness_repr.t -> timestamp:Tezos_protocol_environment_019_PtParisB.Time.t -> delegate_pk:Tezos_protocol_environment_019_PtParisB.Signature.public_key -> round_durations:Round_repr.Durations.t -> proof_of_work_threshold:int64 -> expected_commitment:bool -> unit Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult