Module Carbonated_map.Make_builder

A functor for exposing the type of a carbonated map before the carbonated make is created. This is useful in scenarios where the map that will need to be carbonated is defined before the gas consuming functions for the carbonation are available. See for example Raw_context.

Parameters

module C : COMPARABLE

Signature

type 'a t

module Make (G : GAS) : S with type key = C.t and type context = G.context and type 'a t := 'a t