Module Tezos_raw_protocol_019_PtParisB.Context_binary_proof

type Tezos_protocol_environment_019_PtParisB.Error_monad.error +=

| Expected_binary_proof

val check_is_binary : Tezos_protocol_environment_019_PtParisB.Context.Proof.tree Tezos_protocol_environment_019_PtParisB.Context.Proof.t -> unit Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

is_binary proof will fail if proof does not claim to be a Merkle proof of a binary tree compatible with the one defined in Context_binary.

Note: It is very important to systematically check this before calling the verify_proof exposed in the protocol, since this function does not discriminate between binary and 32-ary proofs.