Module History.Internal_for_tests

type cell_content =

| Unattested of Header.id

| Attested of Header.t

The content of a cell in the DAL skip list. We don't store the slot headers directly to refactor the common published_level and save space. This is important for refutation proofs, as they have to fit in an L1 operation.

val content : t -> cell_content

Returns the content of the last cell in the given skip list.

val proof_statement_is : proof -> [ `Confirmed | `Unconfirmed ] -> bool

proof_statement_is serialized_proof expected will return true if the deserialized proof and the expected proof shape match and false otherwise. Note that it will also return false if deserialization fails.