Module Tezos_raw_protocol_019_PtParisB.Delegate_missed_attestations_storage

This modules deals with delegates' participation in consensus.

This module is responsible for maintaining the Storage.Contract.Missed_attestations table.

val expected_slots_for_given_active_stake : Raw_context.t -> total_active_stake_weight:int64 -> active_stake_weight:int64 -> int

type level_participation =

| Participated

| Didn't_participate

val record_attesting_participation : Raw_context.t -> delegate:Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t -> participation:level_participation -> attesting_power:int -> Raw_context.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

Record the participation of a delegate as a validator.

val record_baking_activity_and_pay_rewards_and_fees : Raw_context.t -> payload_producer: Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t -> block_producer: Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t -> baking_reward:Tez_repr.t -> reward_bonus:Tez_repr.t option -> (Raw_context.t * Receipt_repr.balance_updates) Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

Sets the payload and block producer as active. Pays the baking reward and the fees to the payload producer and the reward bonus to the payload producer (if the reward_bonus is not None).

val check_and_reset_delegate_participation : Raw_context.t -> Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t -> (Raw_context.t * bool) Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

Check that a delegate participated enough in the last cycle (returns true if it did), and then reset the participation for preparing the next cycle.

module For_RPC : sig ... end