Module Tezos_raw_protocol_019_PtParisB.Delegate_storage

This module groups everything related to delegate registration. For the invariants maintained, see the submodule Contract.

It also groups "trivial" getters/setters related to delegates.

It is responsible for maintaining the following tables:

Storage.Contract.Frozen_deposits_limit

Storage.Delegates

type Tezos_protocol_environment_019_PtParisB.Error_monad.error +=

| Unregistered_delegate of Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t

module Contract : sig ... end

This module ensures the following invariants:

val registered : Raw_context.t -> Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t -> bool Tezos_protocol_environment_019_PtParisB.Lwt.t

Has a delegate been registered in the delegate table?

val fold : Raw_context.t -> order:[ `Sorted | `Undefined ] -> init:'a -> f: (Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t -> 'a -> 'a Tezos_protocol_environment_019_PtParisB.Lwt.t) -> 'a Tezos_protocol_environment_019_PtParisB.Lwt.t

Iterate on all registered delegates.

val list : Raw_context.t -> Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t list Tezos_protocol_environment_019_PtParisB.Lwt.t

List all registered delegates.

val initial_frozen_deposits : Raw_context.t -> Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash -> Tez_repr.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

Returns a delegate's initial frozen deposits at the beginning of cycle.

val initial_frozen_deposits_of_previous_cycle : Raw_context.t -> Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash -> Tez_repr.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

Returns a delegate's initial frozen deposits at the beginning of the previous cycle.

Fails with No_previous_cycle if there is no previous cycle.

val current_frozen_deposits : Raw_context.t -> Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash -> Tez_repr.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

Returns a delegate's current frozen deposits, which is the sum of their own frozen funds and those of their stakers if applicable.

val frozen_deposits_limit : Raw_context.t -> Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t -> Tez_repr.t option Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val set_frozen_deposits_limit : Raw_context.t -> Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t -> Tez_repr.t option -> Raw_context.t Tezos_protocol_environment_019_PtParisB.Lwt.t

val spendable_balance : Raw_context.t -> Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash -> Tez_repr.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val drain : Raw_context.t -> delegate:Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t -> destination: Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t -> (Raw_context.t * bool * Tez_repr.t * Receipt_repr.balance_updates) Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

module For_RPC : sig ... end

The functions in this module are considered too costly to be used in the protocol. They are meant to be used only to answer RPC calls.