Module Tezos_raw_protocol_019_PtParisB.Denunciations_repr

type item = {

operation_hash : Tezos_protocol_environment_019_PtParisB.Operation_hash.t;

rewarded : Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash;

misbehaviour : Misbehaviour_repr.t;

}

Internal representation of a pending denunciation, meaning that a denunciation operation has been observed in an applied block, but the corresponding slashing has not happened yet.

Note: the public key hash of the culprit doesn't appear in this type because it is used as key to store the list of a culprit's items (see type t below) in the context.

type t = item list

List of all pending denunciations about the same culprit.

val item_encoding : item Tezos_protocol_environment_019_PtParisB.Data_encoding.t

val encoding : t Tezos_protocol_environment_019_PtParisB.Data_encoding.t

val add : Tezos_protocol_environment_019_PtParisB.Operation_hash.t -> Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash -> Misbehaviour_repr.t -> t -> t

Append a new pending denunciation to the end of the given list.