Module Tezos_raw_protocol_019_PtParisB.Fitness_repr

type Tezos_protocol_environment_019_PtParisB.Error_monad.error +=

| Invalid_fitness

| Wrong_fitness

| Outdated_fitness

| Locked_round_not_less_than_round of {

round : Round_repr.t;

locked_round : Round_repr.t;

}

type t

val encoding : t Tezos_protocol_environment_019_PtParisB.Data_encoding.t

val pp : Tezos_protocol_environment_019_PtParisB.Format.formatter -> t -> unit

val create : level:Raw_level_repr.t -> locked_round:Round_repr.t option -> predecessor_round:Round_repr.t -> round:Round_repr.t -> t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

val create_without_locked_round : level:Raw_level_repr.t -> predecessor_round:Round_repr.t -> round:Round_repr.t -> t

val to_raw : t -> Tezos_protocol_environment_019_PtParisB.Fitness.t

val from_raw : Tezos_protocol_environment_019_PtParisB.Fitness.t -> t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

Returns the corresponding protocol fitness if the shell fitness has the expected version, given by Constants_repr.fitness_version_number. If the fitness' version is from a previous protocol version, then it raises an "outdated fitness" error. If the fitness version is higher then it raises an "invalid fitness" error.

val round_from_raw : Tezos_protocol_environment_019_PtParisB.Fitness.t -> Round_repr.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

Returns the round from a raw fitness. If the fitness is from a previous protocol, the returned value will be Round.zero.

val predecessor_round_from_raw : Tezos_protocol_environment_019_PtParisB.Fitness.t -> Round_repr.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

Returns the predecessor round from a raw fitness. If the fitness is from a previous protocol, the returned value will be Round.zero.

val locked_round_from_raw : Tezos_protocol_environment_019_PtParisB.Fitness.t -> Round_repr.t option Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

Returns the locked round from a raw fitness. If the fitness is from a previous version, the returned value will be None.

val check_except_locked_round : t -> level:Raw_level_repr.t -> predecessor_round:Round_repr.t -> unit Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

Validate only the part of the fitness for which information are available during begin_application

val level : t -> Raw_level_repr.t

val round : t -> Round_repr.t

val locked_round : t -> Round_repr.t option

val predecessor_round : t -> Round_repr.t