Module Tezos_raw_protocol_019_PtParisB.Origination_nonce

type t = {

operation_hash : Tezos_protocol_environment_019_PtParisB.Operation_hash.t;

origination_index : int32;

}

Originated contracts and tx rollups handles are crafted from the hash of the operation that triggered their origination (and nothing else). As a single operation can trigger several originations, the corresponding handles are forged from a deterministic sequence of nonces, initialized with the hash of the operation.

val encoding : t Tezos_protocol_environment_019_PtParisB.Data_encoding.t

val initial : Tezos_protocol_environment_019_PtParisB.Operation_hash.t -> t

val incr : t -> t