Module Tezos_raw_protocol_019_PtParisB.Pending_denunciations_storage

This module deals with pending denunciations before they are used to slash delegates.

This module is responsible for maintaining the table Storage.Pending_denunciations

In particular, it maintains the invariant that no key is pointing to an empty denunciation list.

val find : Raw_context.t -> Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash -> Denunciations_repr.item list Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

Returns the pending denunciations list of the given delegate. It returns an empty list if none are registered.

val add_denunciation : Raw_context.t -> misbehaving_delegate: Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash -> Tezos_protocol_environment_019_PtParisB.Operation_hash.t -> rewarded_delegate: Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash -> Misbehaviour_repr.t -> Raw_context.t Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

Add a denunciation in the list of the given delegate

val set_denunciations : Raw_context.t -> Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash -> Denunciations_repr.t -> Raw_context.t Tezos_protocol_environment_019_PtParisB.Lwt.t

Set the denunciation list of the given delegate. Previously set denunciations would be erased.

val has_pending_denunciations : Raw_context.t -> Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash -> bool Tezos_protocol_environment_019_PtParisB.Lwt.t

Tells if the given delegate has some pending denunciations

val fold : Raw_context.t -> order:[ `Sorted | `Undefined ] -> init:'a -> f: (Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash -> Denunciations_repr.item list -> 'a -> 'a Tezos_protocol_environment_019_PtParisB.Lwt.t) -> 'a Tezos_protocol_environment_019_PtParisB.Lwt.t

See Storage.Pending_denunciations.fold

val clear : Raw_context.t -> Raw_context.t Tezos_protocol_environment_019_PtParisB.Lwt.t

See Storage.Pending_denunciations.clear

module For_RPC : sig ... end