Module Sc_rollup_game_repr.Internal_for_tests

val find_choice : dissection_chunk list -> Sc_rollup_tick_repr.t -> (dissection_chunk * dissection_chunk) Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

Checks that the tick count chosen by the current move is one of the ones in the current dissection. Returns a tuple containing the current dissection interval (including the two states) between this tick and the next.

val check_dissection : default_number_of_sections:int -> start_chunk:dissection_chunk -> stop_chunk:dissection_chunk -> dissection_chunk list -> unit Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

See Sc_rollup_dissection_chunk_repr.default_check