Module Tezos_raw_protocol_019_PtParisB.Sc_rollup_management_protocol

This module provides a typed API for the Rollup Management Protocol that defines the communication protocol for exchanging messages between Layer 1 and Layer 2 for smart-contract rollups.

The API exposes functions for constructing inbox messages. These are messages produced by the Layer 1 protocol and added to a smart-contract rollups inbox.

The Layer 2 node is responsible for decoding and interpreting the messages.

A type outbox_message representing messages from Layer 2 to Layer 1 is also provided. An outbox_message consists of a set of transactions to L1 accounts.

type Tezos_protocol_environment_019_PtParisB.Error_monad.error +=

| Sc_rollup_invalid_destination

type transaction = private

| Transaction : {

destination : Contract_hash.t;

entrypoint : Alpha_context.Entrypoint.t;

parameters_ty : ('a, _) Script_typed_ir.ty;

parameters : 'a;

unparsed_parameters : Alpha_context.Script.expr;

} -> transaction

A type representing a Layer 2 to Layer 1 transaction.

type atomic_transaction_batch = private {

transactions : transaction list;

}

A type representing a batch of Layer 2 to Layer 1 transactions.

type outbox_message = private

| Atomic_transaction_batch of atomic_transaction_batch

| Whitelist_update of Alpha_context.Sc_rollup.Whitelist.t option

A typed representation of Sc_rollup.Outbox.Message.t.

val make_internal_transfer : Alpha_context.context -> ('a, _) Script_typed_ir.ty -> payload:'a -> sender:Contract_hash.t -> source:Alpha_context.public_key_hash -> destination:Alpha_context.Sc_rollup.Address.t -> (Alpha_context.Sc_rollup.Inbox_message.t * Alpha_context.context) Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

make_internal_transfer ctxt ty ~payload ~sender ~source ~destination constructs a smart rollup's inbox message (an L1 to L2 message) with the given payload, sender, and source targeting destination.

val outbox_message_of_outbox_message_repr : Alpha_context.context -> Alpha_context.Sc_rollup.Outbox.Message.t -> (outbox_message * Alpha_context.context) Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

outbox_message_of_outbox_message_repr ctxt msg returns a typed version of of the given outbox message msg.

Fails with an Sc_rollup_invalid_destination error in case the parameters don't match the type of the entrypoint and destination.

module Internal_for_tests : sig ... end

Function for constructing and encoding inbox_message and outbox_message values. Since Layer 1 only ever consumes outbox_message values and produces inbox_message values, these functions are used for testing only.