Module Storage.Slashed_deposits_history

type slashed_percentage = Percentage.t

type t = (Cycle_repr.t * slashed_percentage) list

val add : Cycle_repr.t -> slashed_percentage -> t -> t

add cycle percentage history adds the percentage for the cycle in the history. If the cycle exists, the associated percentage is updated and capped at 100 and the cycle order in the list is unchanged. If the cycle does not exist, the new pair (cycle, percentage) is added at the beginning of the list.

val get : Cycle_repr.t -> t -> slashed_percentage

get cycle history returns the percentage for cycle in history or 0 if there is no such cycle.