Module Storage.Stake

module Staking_balance : Storage_sigs.Indexed_data_storage with type key = Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t and type value = Full_staking_balance_repr.t and type t := Raw_context.t

The map of all the stake of all delegates, including those with less than Constants_parametric_repr.t.minimal_stake. It might be large.

module Active_delegates_with_minimal_stake : Storage_sigs.Data_set_storage with type elt = Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t and type t := Raw_context.t

This should be fairly small compared to staking balance

module Selected_distribution_for_cycle : Storage_sigs.Indexed_data_storage with type key = Cycle_repr.t and type value = (Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t * Stake_repr.t) list and type t := Raw_context.t

List of active stake

module Total_active_stake : Storage_sigs.Indexed_data_storage with type key = Cycle_repr.t and type value = Stake_repr.t and type t := Raw_context.t

Sum of the active stakes of all the delegates with Constants_parametric_repr.t.minimal_stake