Module Tezos_raw_protocol_019_PtParisB.Storage_costs

val read_access : path_length:int -> read_bytes:int -> Gas_limit_repr.cost

Cost of reading read_bytes at a key of length path_length.

val write_access : written_bytes:int -> Gas_limit_repr.cost

Cost of performing a single write access, writing written_bytes bytes.

val list_key_values_traverse : size:int -> Gas_limit_repr.cost

list_key_values_traverse ~size returns the cost of traversing a context with size number of elements.