Module Tezos_raw_protocol_019_PtParisB.Ticket_receipt

A module for representing the increase/decrease of tickets in the storage. It will be used to display ticket update information in the operation receipt.

type update = {

account : Alpha_context.Destination.t;

amount : Tezos_protocol_environment_019_PtParisB.Z.t;

}

Represents that account's storage has delta amount for a given ticket

type item = {

ticket_token : Ticket_token.unparsed_token;

updates : update list;

}

List of updates for a ticket

type t = item list

A list of ticket tokens and their corresponding updates

val item_encoding : item Tezos_protocol_environment_019_PtParisB.Data_encoding.t

val encoding : t Tezos_protocol_environment_019_PtParisB.Data_encoding.t