Module Tezos_raw_protocol_019_PtParisB.Unstaked_frozen_staker_repr

type t =

| Single of Contract_repr.t * Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash
(*
A single staker, either the delegate itself or one of its staker.
*)

| Shared of Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash
(*
The delegate and all its stakers simultaneously.
*)

Adding and removing unstaked frozen stake can be done from/toward a delegate, one of its staker, or both the delegate and all its stakers at once. We need to distinguish these cases to enforce the staking over baking limit.

val encoding : t Tezos_protocol_environment_019_PtParisB.Data_encoding.t

val compare : t -> t -> int

val delegate : t -> Tezos_protocol_environment_019_PtParisB.Signature.public_key_hash