Module Tezos_raw_protocol_019_PtParisB.Zk_rollup_update_repr

Payload of a ZK Rollup update operation. The operator only needs to send a subset of the public inputs defined in Zk_rollup_circuit_public_inputs_repr, the rest is provided by the protocol.

type op_pi = {

new_state : Zk_rollup_state_repr.t;

fee : Zk_rollup_scalar.t;

exit_validity : bool;

}

Minimal subset of public inputs for the public L2 operations' circuits.

type private_inner_pi = {

new_state : Zk_rollup_state_repr.t;

fees : Zk_rollup_scalar.t;

}

Minimal subset of public inputs for the circuits for batches of private L2 operations

type fee_pi = {

new_state : Zk_rollup_state_repr.t;

}

Minimal subset of public inputs for the "fee" circuit.

type t = {

pending_pis : (string * op_pi) list;

private_pis : (string * private_inner_pi) list;

fee_pi : fee_pi;

proof : Tezos_protocol_environment_019_PtParisB.Plonk.proof;

}

Payload of an update operation. Includes the proof and the public inputs that are needed to verify it. Each set of public inputs also carries the string that identifies the circuit which they are for.

val encoding : t Tezos_protocol_environment_019_PtParisB.Data_encoding.t